segundo-Extra-commit-task-11

+++ src/main/java/com/mindhub/homebanking/dtos/LoanAplicationDTO.java
@@ -6,28 +6,31 @@ import java.util.Set;
 public class LoanAplicationDTO {
     public Long id;
 
-    public long amount;
+    public double amount;
 
     private Integer payments;
 
     public String number;
 
+
     public LoanAplicationDTO() {
 
     }
 
-    public LoanAplicationDTO(Long id, long amount, Integer payments, String number) {
+    public LoanAplicationDTO(Long id, double amount, Integer payments, String number) {
         this.id = id;
         this.amount = amount;
         this.payments = payments;
         this.number = number;
+
     }
+
 
     public Long getId() {
         return id;
     }
 
-    public long getAmount() {
+    public double getAmount() {
         return amount;
     }
